requirements.txt
asgiref==3.5.2
autopep8==1.7.0
Brotli==1.0.9
dj-database-url==1.0.0
Django==4
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.4
pycodestyle==2.9.1
pytz==2022.4
South==1.0.2
sqlparse==0.4.3
toml==0.10.2
tzdata==2022.4
whitenoise==6.2.0
